Refactor: replace tsdrg implemented with google.tensornetwork to quimb
