+++ tnpy/tsdrg.py
@@ -1,112 +1,117 @@
 import logging
-import numpy as np
-from functools import wraps
-from itertools import count
-from collections import deque
-from copy import copy
-from graphviz import Digraph
-from tensornetwork import ncon, Node, Tensor
+import re
 from dataclasses import dataclass, field
-from tnpy.operators import MPO
-from typing import Tuple, List, Union, Callable, Iterator, Dict
+from functools import wraps
+from itertools import count, islice
+from typing import Union, List, Dict, Iterator, Callable, Sequence, Tuple
+
+import numpy as np
+import quimb.tensor as qtn
+from graphviz import Digraph
+
+from tnpy.operators import MatrixProductOperator
 
 
-@dataclass
-class TreeNode:
-    """
-    The node object in binary tree.
+class Node(qtn.Tensor):
 
-    Attributes:
-        id: ID of this TreeNode.
-        node: Container of :class:`~tensornetwork.Node`,
-            which is supposed to take the ordering of axes `(1st physical bond, 2nd physical bond, virtual bond)`.
-        left:
-        right:
-
-    Notes:
-        This class differs from :class:`~tensornetwork.Node`,
-        which refers to the node in a tensor network.
-        Please be aware of this bad naming.
-    """
-    id: int
-    node: Node
-    left: 'TreeNode' = None
-    right: 'TreeNode' = None
-
-    def __eq__(self, other: 'TreeNode') -> bool:
+    def __init__(self, node_id: int, left: 'Node' = None, right: 'Node' = None,
+                 *args, **kwargs):
         """
-        Check 2 TreeNodes are identical.
+        The node of binary tree, while data structure is inherited from :class:`~quimb.Tensor`.
 
         Args:
-            other:
-
-        Returns:
-
-        Warnings:
-            This is a weak condition, only the id of these 2 TreeNodes
-            and the id of their 1st generation children (if not a leaf) are compared.
-            It's not guaranteed that the value of :attr:`~TreeNode.node` will be the same,
-            as well as their posterity.
+            node_id:
+            left:
+            right:
+            *args:
+            **kwargs:
         """
-        if self.is_leaf and other.is_leaf:
-            return self.id == other.id
-        elif self.is_leaf or other.is_leaf:
-            return False
-        return self.id == other.id and self.left.id == other.left.id and self.right.id == other.right.id
+        super(Node, self).__init__(*args, **kwargs)
+        self._node_id = node_id
+        self._left = left
+        self._right = right
+
+    def __repr__(self):
+        left_node_id = None if self.left is None else self.left.node_id
+        right_node_id = None if self.right is None else self.right.node_id
+        return f'Node(id={self.node_id}, ' \
+               f'left_id={left_node_id}, ' \
+               f'right_id={right_node_id}, ' \
+               f'shape={tuple(map(int, self.data.shape))}, ' \
+               f'inds={self.inds})'
+
+    @property
+    def node_id(self) -> int:
+        return self._node_id
+
+    @property
+    def left(self) -> 'Node':
+        return self._left
+
+    @property
+    def right(self) -> 'Node':
+        return self._right
 
     @property
     def is_leaf(self) -> bool:
         return self.left is None and self.right is None
 
-    @property
-    def shape(self) -> Tuple[int]:
-        return self.node.tensor.shape
-
-
-def check_root(func: Callable) -> Callable:
-    """
-    Helper function for checking the root in :class:`~TensorTree`.
-    This is supposed to be applied on the member function of :class:`~TensorTree` as a decorator.
-
-    Args:
-        func:
-
-    Returns:
-
-    """
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        assert isinstance(args[0], TensorTree), "This is supposed to be used on class member of TensorTree."
-        if not args[0].has_root:
-            raise RuntimeError("Cannot find root in tree.")
-        return func(*args, **kwargs)
-    return wrapper
-
 
 class TensorTree:
 
-    def __init__(self, leaves: List[TreeNode]):
+    @dataclass
+    class Syntax:
+        node: str = 'Node'
+        conj_node: str = 'ConjNode'
+        level_idx: str = 'LevelIdx'
+
+    def __init__(self, mpo: MatrixProductOperator):
         """
         The bottom-up binary tree where each node contains a tensor.
 
         Args:
-            leaves:
+            mpo:
         """
-        self._leaves = {leaf.id: leaf for leaf in leaves}
-        self._tree = copy(self._leaves)
-        self._horizon = list(self._tree.keys())
-        self._root = None
+        self._root_id = None
+        self._tree = {
+            site: Node(
+                node_id=site,
+                data=tensor.data,
+                inds=tensor.inds,
+                tags=tensor.tags
+            )
+            for site, tensor in enumerate(mpo)
+        }
+        self._n_leaves = mpo.nsites
+        self._horizon = list(range(self._n_leaves))
 
     def __iter__(self) -> Iterator:
         return iter(self._tree.values())
 
-    @property
-    def keys(self) -> List:
-        return list(self._tree.keys())
+    def __getitem__(self, node_id: int) -> Node:
+        return self._tree[node_id]
+
+    def __setitem__(self, node_id: int, node: Node):
+        self._tree[node_id] = node
+
+    def __repr__(self):
+        return 'TensorTree([\n\t{}\n])'.format(', \n\t'.join(repr(node) for node in self._tree.values()))
 
     @property
-    def leaves(self) -> Dict:
-        return self._leaves
+    def root_id(self) -> int:
+        return self._root_id
+
+    @property
+    def root(self) -> Node:
+        return self[self._root_id]
+
+    @property
+    def leaves(self) -> Dict[int, Node]:
+        return dict(islice(self._tree.items(), self._n_leaves))
+
+    @property
+    def horizon(self) -> List[int]:
+        return self._horizon
 
     @property
     def n_nodes(self) -> int:
@@ -119,8 +124,55 @@ class TensorTree:
         return len(self._tree)
 
     @property
+    def n_leaves(self) -> int:
+        return self._n_leaves
+
+    def fuse(self, left_id: int, right_id: int, new_id: int, data: np.ndarray):
+        if self.root_id is not None:
+            raise RuntimeError("Can't perform fuse on a grown tree.")
+        left_ind = f'{self.Syntax.node}{new_id}-{self.Syntax.node}{left_id}' \
+            if not self[left_id].is_leaf else f'k{left_id}'
+        right_ind = f'{self.Syntax.node}{new_id}-{self.Syntax.node}{right_id}' \
+            if not self[right_id].is_leaf else f'k{right_id}'
+        self[new_id] = Node(
+            node_id=new_id,
+            left=self[left_id],
+            right=self[right_id],
+            data=data,
+            inds=(left_ind, right_ind, f'{self.Syntax.node}{new_id}{self.Syntax.level_idx}'),
+            tags=f'{self.Syntax.node}{new_id}'
+        )
+        self[left_id].reindex(
+            {f'{self.Syntax.node}{left_id}{self.Syntax.level_idx}':
+                f'{self.Syntax.node}{new_id}-{self.Syntax.node}{left_id}'}, inplace=True
+        )
+        self[right_id].reindex(
+            {f'{self.Syntax.node}{right_id}{self.Syntax.level_idx}':
+                f'{self.Syntax.node}{new_id}-{self.Syntax.node}{right_id}'}, inplace=True
+        )
+        self._horizon[self._horizon.index(left_id)] = new_id
+        self._horizon.remove(right_id)
+        if self.n_nodes == 2 * self.n_leaves - 1:
+            self._root_id = new_id
+
+    def check_root(func: Callable) -> Callable:
+        """
+        Accessory decorator for checking the existence of root in :class:`~TensorTree`.
+
+        Returns:
+
+        """
+        @wraps(func)
+        def wrapper(self, *args, **kwargs):
+            if self._root_id is None:
+                raise KeyError("Can't find root in tree.")
+            return func(self, *args, **kwargs)
+        return wrapper
+
+    @property
+    @check_root
     def n_layers(self) -> int:
-        def max_depth(current_node: TreeNode) -> int:
+        def max_depth(current_node: Node) -> int:
             if current_node is not None:
                 left_depth = max_depth(current_node.left)
                 right_depth = max_depth(current_node.right)
@@ -128,69 +180,61 @@ class TensorTree:
             return 0
         return max_depth(self.root)
 
-    @property
-    def n_leaves(self) -> int:
-        return len(self._leaves)
-
-    @property
-    def root(self) -> Union[None, TreeNode]:
-        return self._root
-
-    @property
-    def has_root(self) -> bool:
-        return self.root is not None
-
-    @property
-    def horizon(self) -> List[int]:
-        return self._horizon
-
-    def append(self, node: TreeNode) -> None:
-        self._tree[node.id] = node
-        self._horizon[self._horizon.index(node.left.id)] = node.id
-        self._horizon.remove(node.right.id)
-        if self.n_nodes == 2 * self.n_leaves - 1:
-            self._root = node
-
-    def node(self, node_id: int) -> TreeNode:
-        return self._tree[node_id]
-
-    def leaf(self, node_id: int) -> TreeNode:
-        return self._leaves[node_id]
+    @check_root
+    def tensor_network(self, node_ids: Sequence[int] = None, conj: bool = False,
+                       mangle_outer: bool = True, with_leaves: bool = False) -> qtn.TensorNetwork:
+        node_ids = list(self._tree.keys()) if node_ids is None else node_ids
+        nodes = [self[node_id] for node_id in node_ids if not self[node_id].is_leaf] if not with_leaves \
+            else [self[node_id] for node_id in node_ids]
+        net = qtn.TensorNetwork(nodes)
+        if conj:
+            conj_net = net.reindex(
+                {f'{self.Syntax.node}{self._root_id}{self.Syntax.level_idx}':
+                    f'{self.Syntax.conj_node}{self._root_id}{self.Syntax.level_idx}'}
+            )
+            conj_net = conj_net.retag(
+                {tag: tag.replace(self.Syntax.node, self.Syntax.conj_node)
+                 for node in conj_net for tag in node.tags}
+            )
+            if mangle_outer:
+                conj_net = conj_net.reindex(
+                    {ind: re.sub(r'^k([0-9]+)', r'b\1', ind)
+                     for node in conj_net for ind in node.inds if ind.startswith('k')}
+                )
+            return conj_net
+        return net
 
     @check_root
-    def find_path(self, node_id: int) -> List[int]:
+    def find_path(self, node_id: int, return_itself: bool = False) -> List[int]:
         """
         Find the path to targeted node starting from the root.
 
         Args:
             node_id:
+            return_itself:
 
         Returns:
             path: ID of nodes to targeted node, including targeted node itself.
 
         Raises:
             If path is empty, raises KeyError for input ``node_id``.
-
-        See Also:
-            The function :py:func:`TensorTree.ancestor` does the same job as this function,
-            while the returning type is List[:class:`~TreeNode`] there.
         """
-        def find_node(current_node: TreeNode, trial_path: List[int]) -> bool:
+        def find_node(current_node: Node, trial_path: List[int]) -> bool:
             """
             Find the node with depth-first search.
 
             Args:
                 current_node: Current node in iterative depth-first searching
-                trial_path: List for recording the trial path to every leaf.
+                trial_path: List for recording the trial path to the targeted leaf.
 
             Returns:
                 found: Return True if the given ``node_id`` is found in tree, else False.
             """
             if current_node is not None:
-                trial_path.append(current_node.id)
-                if current_node.id == node_id:
+                trial_path.append(current_node.node_id)
+                if current_node.node_id == node_id:
                     return True
-                elif find_node(current_node.left, trial_path) or\
+                elif find_node(current_node.left, trial_path) or \
                         find_node(current_node.right, trial_path):
                     return True
                 trial_path.pop()
@@ -199,173 +243,40 @@ class TensorTree:
         path = []
         if not find_node(self.root, path):
             raise KeyError("The given node_id is not found in tree.")
-        return path
-
-    def ancestor(self, node_id: int) -> List[TreeNode]:
-        """
-        Find all the ancestors of targeted node in the order starting from the root.
-
-        Args:
-            node_id:
-
-        Returns:
-            ancestor:
-
-        See Also:
-            The function :func:`~TensorTree.find_path` does the same job as this function,
-            while the returning type is List[int] there.
-        """
-        path = self.find_path(node_id)
-        return [self.node(idx) for idx in path][:-1]
+        return path if return_itself else path[:-1]
 
     @check_root
-    def common_ancestor(self, node_id1: int, node_id2: int) -> List[int]:
+    def common_ancestor(self, node_id1: int, node_id2: int, lowest: bool = False) -> Union[int, List[int]]:
         """
         Find all common ancestor of two given nodes in the order starting from the root.
 
         Args:
             node_id1:
             node_id2:
+            lowest:
 
         Returns:
 
         """
-        path1 = self.find_path(node_id1)
-        path2 = self.find_path(node_id2)
+        path1 = self.find_path(node_id1, return_itself=True)
+        path2 = self.find_path(node_id2, return_itself=True)
         try:
             first_unmatched = next(idx for idx, (x, y) in enumerate(zip(path1, path2)) if x != y)
-            return path1[:first_unmatched]
+            common_path = path1[:first_unmatched]
         except StopIteration:
             assert path1 == path2
-            return path1[:-1]
-
-    def lowest_common_ancestor(self, node_id1: int, node_id2: int) -> int:
-        return self.common_ancestor(node_id1, node_id2)[-1]
-
-    def filter_out_leaf(self, node_ids: List[int]) -> List[int]:
-        return [node_id for node_id in node_ids if not self.node(node_id).is_leaf]
-
-    class Contractible:
-        """
-        Accessory helper class for contracting TreeNodes.
-        """
-
-        def __init__(self, parent_cls: 'TensorTree', node_ids: List[int], open_bonds: List[Tuple[int, str]]):
-            self.parent_cls = parent_cls
-            self.open_bonds = open_bonds
-            self.input_checker(node_ids, open_bonds)
-            self._contractible = {
-                k: v for node_id in node_ids for k, v in zip(
-                    [f"{node_id}", f"conj{node_id}"],
-                    [self.cell(node_id), self.cell(node_id, conj=True)]
-                )
-            }
-
-        def __getitem__(self, node_id: str):
-            return self._contractible[node_id]['bonds']
-
-        @property
-        def tensors(self) -> List[Tensor]:
-            return [node['tensor'] for node in self._contractible.values()]
-
-        @property
-        def network_structure(self) -> List[Tuple[str, str, str]]:
-            return [tuple(node['bonds']) for node in self._contractible.values()]
-
-        @property
-        def out_order(self) -> List[str]:
-            edge_bonds = [self.syntax(*open_bond) for open_bond in self.open_bonds]
-            conj_edge_bonds = [self.syntax(*open_bond, conj=True) for open_bond in self.open_bonds]
-            return [self.syntax(self.parent_cls.root.id, 2)] + edge_bonds +\
-                   [self.syntax(self.parent_cls.root.id, 2, conj=True)] + conj_edge_bonds
-
-        @staticmethod
-        def syntax(node_id: int, idx: Union[int, str], conj: bool = False) -> str:
-            if type(idx) is str:
-                idx = 0 if idx == 'left' else 1
-            prefix = "Conj" if conj else ""
-            return f"-{prefix}Node{node_id}Sub{idx}"
-
-        def cell(self, node_id: int, conj: bool = False) -> Dict:
-            def _tensor_setter() -> Tensor:
-                node = self.parent_cls.node(node_id).node
-                return node.copy().tensor if conj else node.tensor
-
-            def _bonds_setter() -> List[str]:
-                return [self.syntax(node_id, idx, conj) for idx in range(3)]
-            return {'tensor': _tensor_setter(), 'bonds': _bonds_setter()}
-
-        def input_checker(self, node_ids: List[int], open_bonds: List[Tuple[int, str]]):
-            if len(node_ids) > len(set(node_ids)):
-                raise KeyError("Every id in node_ids must be given uniquely.")
-            for elem in open_bonds:
-                if not isinstance(elem, tuple) and list(map(type, elem)) == [int, str]:
-                    raise TypeError("Argument `open_bonds` takes the type List[Tuple[int, str]].")
-                if elem[1] not in ['left', 'right']:
-                    raise KeyError("Second argument of `open_bonds` must be either `left` or `right`.")
-                if getattr(self.parent_cls.node(elem[0]), elem[1]).id in node_ids:
-                    raise KeyError("Child node of `open_bonds` cannot present in `node_ids`.")
+            common_path = path1[:-1]
+        return common_path if not lowest else common_path[-1]
 
     @check_root
-    def contract_nodes(self, node_ids: List[int], open_bonds: List[Tuple[int, str]] = None,
-                       return_out_order: bool = False) -> Union[Tensor, Tuple[Tensor, List[str]]]:
-        """
-
-        Args:
-            node_ids:
-            open_bonds:
-            return_out_order:
-
-        Returns:
-            contracted_network:
-            out_order:
-        """
-        if open_bonds is None:
-            open_bonds = []
-        node_queue = deque([self.root])
-        contractible = self.Contractible(self, node_ids, open_bonds)
-
-        def update_contractible(parent: TreeNode, on: str):
-            child = getattr(parent, on)
-            on_bond_idx = 0 if on == 'left' else 1
-            if child.id in node_ids:
-                assert not child.is_leaf, "node_ids contains leaves"
-                contractible[f"{parent.id}"][on_bond_idx] = f"Node{parent.id}ToNode{child.id}"
-                contractible[f"{child.id}"][2] = f"Node{parent.id}ToNode{child.id}"
-                contractible[f"conj{parent.id}"][on_bond_idx] = f"ConjNode{parent.id}ToConjNode{child.id}"
-                contractible[f"conj{child.id}"][2] = f"ConjNode{parent.id}ToConjNode{child.id}"
-                if (child.id, 'left') or (child.id, 'right') not in open_bonds:
-                    node_queue.append(child)
-            elif (parent.id, on) not in open_bonds:
-                contractible[f"conj{parent.id}"][on_bond_idx] = f"Node{parent.id}Sub{on_bond_idx}" \
-                                                                f"ToConjNode{parent.id}Sub{on_bond_idx}"
-                contractible[f"{parent.id}"][on_bond_idx] = f"Node{parent.id}Sub{on_bond_idx}" \
-                                                            f"ToConjNode{parent.id}Sub{on_bond_idx}"
-
-        while len(node_queue) > 0:
-            current_node = node_queue.popleft()
-            update_contractible(current_node, on='left')
-            update_contractible(current_node, on='right')
-
-        # TODO: con_order should be given
-        out_tensor = ncon(
-            contractible.tensors,
-            contractible.network_structure,
-            out_order=contractible.out_order
-        )
-        if return_out_order:
-            return out_tensor, contractible.out_order
-        return out_tensor
-
-    @check_root
-    def plot(self) -> Digraph:
-        def find_child(node: TreeNode):
+    def plot(self, view: bool = False) -> Digraph:
+        def find_child(node: Node):
             for attr in ['left', 'right']:
                 if getattr(node, attr) is not None:
                     if not getattr(node, attr).is_leaf:
-                        graph.node(f'{getattr(node, attr).id}', shape='triangle', style='rounded')
+                        graph.node(f'{getattr(node, attr).node_id}', shape='triangle', style='rounded')
                     graph.edge(
-                        f'{node.id}', f'{getattr(node, attr).id}',
+                        f'{node.node_id}', f'{getattr(node, attr).node_id}',
                         splines='ortho', minlen='2',
                         headport='n', tailport='_', arrowhead='inv', constraint='true'
                     )
@@ -373,114 +284,92 @@ class TensorTree:
 
         graph = Digraph()
         graph.node('head', style='invis')
-        graph.edge('head', f'{self.root.id}', headport='n', arrowhead="inv")
-        graph.node(f'{self.root.id}', shape='triangle', rank='max', style='rounded')
+        graph.edge('head', f'{self.root.node_id}', headport='n', arrowhead="inv")
+        graph.node(f'{self.root.node_id}', shape='triangle', rank='max', style='rounded')
         find_child(self.root)
-        with graph.subgraph(name='cluster_0') as sg:
+        with graph.subgraph(name='cluster_0') as sg:  # for constraining mpo inside an invisible box
             sg.attr(style='invis')
-            for k, v in enumerate(self._leaves.keys()):
+            for k, v in enumerate(self.leaves.keys()):
                 sg.node(f'{v}', shape='box', rank='sink', style='rounded')
-                if k < len(self._leaves.keys()) - 1:
-                    sg.edge(f'{v}', f'{v+1}', splines='ortho', minlen='0', arrowhead="none", constraint='true')
+                if k < len(self.leaves.keys()) - 1:
+                    sg.edge(f'{v}', f'{v + 1}', splines='ortho', minlen='0', arrowhead="none", constraint='true')
         logging.debug(graph)
+        if view:
+            graph.render(format='png', view=True)
         return graph
 
 
-class TSDRG:
+class TreeTensorNetworkSDRG:
 
     @dataclass
     class GapCache:
         """
-        Helper class for storing the cache of energy gap in tSDRG algorithm.
+        Helper class for caching the energy gap in tSDRG algorithm.
         """
-        parent_cls: 'TSDRG'
-        gap: List[float] = field(default_factory=list)
+        tsdrg: 'TreeTensorNetworkSDRG'
         evecs: List[np.ndarray] = field(default_factory=list)
-        evals: np.ndarray = None
+        gap: List[float] = field(default_factory=list)
 
         def __post_init__(self):
-            for site in range(self.parent_cls.n_nodes - 1):
-                evals, evecs = self.parent_cls.eigen_solver(self.parent_cls.block_hamiltonian(site))
-                self.gap.append(self.parent_cls.truncation_gap(evals))
+            for site in range(self.tsdrg.n_sites - 1):
+                evals, evecs = self.tsdrg.eigen_solver(self.tsdrg.block_hamiltonian(site))
                 self.evecs.append(evecs)
-                if self.parent_cls.n_nodes == 2:
-                    self.evals = evals
+                self.gap.append(self.tsdrg.truncation_gap(evals))
+                self.tsdrg._evals = evals if self.tsdrg.n_sites == 2 else None
 
-    def __init__(self, mpo: MPO, chi: int):
-        self.mpo = mpo
+        def neighbouring_bonds(self, bond: int) -> List[int]:
+            if bond == 0:
+                neighbours = [bond]
+            elif bond == len(self.tsdrg._fused_mpo_cache) - 1:
+                neighbours = [bond - 1]
+            else:
+                neighbours = [bond - 1, bond]
+            return neighbours
+
+        def update(self, max_gapped_bond: int):
+            self.gap.pop(max_gapped_bond)
+            self.evecs.pop(max_gapped_bond)
+            for bond in self.neighbouring_bonds(max_gapped_bond):
+                evals, evecs = self.tsdrg.eigen_solver(self.tsdrg.block_hamiltonian(bond))
+                self.gap[bond] = self.tsdrg.truncation_gap(evals)
+                self.evecs[bond] = evecs
+                self.tsdrg._evals = evals
+
+    def __init__(self, mpo: MatrixProductOperator, chi: int):
+        self._mpo = mpo
         self._chi = chi
-        self._N = len(mpo.nodes)
-        self._physical_dimensions = mpo.physical_dimensions
-        assert self._N >= 2, "There must be more than 2 sites by definition."
-        self._tree = TensorTree(
-            [
-                TreeNode(id=site, node=node)
-                for site, node in enumerate(self.mpo)
-            ]
-        )
-        self.gap_cache = self.GapCache(self)
+        self._tree = TensorTree(mpo)
+        self._fused_mpo_cache = [tensor for tensor in mpo]
+        self._gap_cache = self.GapCache(self)
+        self._evals = None
 
     @property
-    def N(self) -> int:
-        return self._N
-
-    @property
-    def physical_dimensions(self) -> int:
-        return self._physical_dimensions
-
-    @property
-    def n_nodes(self) -> int:
-        return len(self.mpo.nodes)
+    def mpo(self) -> MatrixProductOperator:
+        return self._mpo
 
     @property
     def chi(self) -> int:
         return self._chi
 
     @property
-    def v_left(self) -> Node:
-        return self.mpo.v_left
-
-    @property
-    def v_right(self) -> Node:
-        return self.mpo.v_right
-
-    @property
     def tree(self) -> TensorTree:
         return self._tree
 
-    def block_hamiltonian(self, site: int) -> Union[Tensor, np.ndarray]:
-        """
-        Construct the 2-site Hamiltonian from MPO,
-        involving both the hopping term and one-site term.
+    @property
+    def n_sites(self) -> int:
+        return self.mpo.nsites
 
-        Args:
-            site: The site i.
-                The 2-site Hamiltonian is then constructed in site i and i+1.
+    @property
+    def evals(self) -> np.ndarray:
+        return self._evals
 
-        Returns:
-            M: The 2-site Hamiltonian.
-        """
-        W1 = self.mpo[site]
-        W2 = self.mpo[site + 1]
-        if self.n_nodes == 2:
-            W1[0] ^ W2[0]
-            M = W1 @ W2
-        elif site == 0:
-            W1[0] ^ W2[0]
-            W2[1] ^ self.v_right[0]
-            M = W1 @ W2 @ self.v_right
-        elif site == self.n_nodes - 2:
-            self.v_left[0] ^ W1[0]
-            W1[1] ^ W2[0]
-            M = self.v_left @ W1 @ W2
-        else:
-            self.v_left[0] ^ W1[0]
-            W1[1] ^ W2[0]
-            W2[1] ^ self.v_right[0]
-            M = self.v_left @ W1 @ W2 @ self.v_right
-        M.reorder_axes([0, 2, 1, 3])
-        shape = int(np.sqrt(M.tensor.size))
-        return M.tensor.reshape(shape, shape)
+    def eigen_solver(self, matrix: np.ndarray) -> Tuple[np.ndarray, np.ndarray]:
+        # TODO: perhaps consider iterative solver?
+        evals, evecs = np.linalg.eigh(matrix)
+        if matrix.shape[0] > self.chi:
+            evals = evals[:self.chi]
+            evecs = evecs[:, :self.chi]
+        return evals, evecs
 
     def truncation_gap(self, evals: np.ndarray) -> float:
         """
@@ -495,161 +384,145 @@ class TSDRG:
         gaps = np.diff(evals)
         return gaps[self.chi - 1] if gaps.size > self.chi else gaps[-1]
 
-    def eigen_solver(self, matrix: Union[Tensor, np.ndarray]) -> Tuple[np.ndarray, np.ndarray]:
-        evals, evecs = np.linalg.eigh(matrix)
-        if matrix.shape[0] > self.chi:
-            evals = evals[:self.chi]
-            evecs = evecs[:, :self.chi]
-        return evals, evecs
-
-    def spectrum_projector(self, site: int, evecs: np.ndarray) -> Tuple[Node, Node]:
+    def block_hamiltonian(self, locus: int) -> np.ndarray:
         """
-
+        Construct the 2-site Hamiltonian from coarse-graining MPO.
 
         Args:
-            site:
+            locus: The site in coarse-grained system during the RG process.
+
+        Returns:
+            block_ham: The 2-site Hamiltonian.
+        """
+        ham = self._fused_mpo_cache[locus] @ self._fused_mpo_cache[locus + 1]
+        if len(self._fused_mpo_cache) > 2:
+            if locus == 0:
+                ham.isel({ham.inds[2]: -1}, inplace=True)
+            elif locus == len(self._fused_mpo_cache) - 2:
+                ham.isel({ham.inds[0]: 0}, inplace=True)
+            else:
+                ham.isel({ham.inds[0]: 0, ham.inds[3]: -1}, inplace=True)
+        ham.fuse({'0': [ham.inds[0], ham.inds[2]], '1': [ham.inds[1], ham.inds[3]]}, inplace=True)
+        return ham.data
+
+    def spectrum_projector(self, locus: int, evecs: np.ndarray) -> np.ndarray:
+        """
+        Coarse-grain the MPO with given evecs which is used to form a projector.
+
+        Args:
+            locus: The site in coarse-grained system during the RG process.
             evecs:
 
         Returns:
-            V: The isometric tensor
-            W: The coarse-grained MPO
+            projector:
         """
-        def network_structure() -> List:
-            if self.n_nodes == 2:
-                return [(1, 'a1', 'a2'), (1, 'b1', 'b2'), ('a1', 'b1', -3), ('a2', 'b2', -4)]
-            elif site == 0:
-                return [(1, 'a1', 'a2'), (1, -2, 'b1', 'b2'), ('a1', 'b1', -3), ('a2', 'b2', -4)]
-            elif site == self.n_nodes - 2:
-                return [(-1, 1, 'a1', 'a2'), (1, 'b1', 'b2'), ('a1', 'b1', -3), ('a2', 'b2', -4)]
-            return [(-1, 1, 'a1', 'a2'), (1, -2, 'b1', 'b2'), ('a1', 'b1', -3), ('a2', 'b2', -4)]
-        W1 = self.mpo[site].tensor
-        W2 = self.mpo[site + 1].tensor
-        V = Node(evecs.reshape((W1.shape[-1], W2.shape[-1], evecs.shape[1])))
-        W = Node(ncon(
-            [W1, W2, V.copy(conjugate=True).tensor, V.tensor],
-            network_structure()
-        ))
-        return V, W
+        mpo1, mpo2 = self._fused_mpo_cache[locus], self._fused_mpo_cache[locus + 1]
+        projector = qtn.Tensor(
+            evecs.reshape((mpo1.shape[-1], mpo2.shape[-1], evecs.shape[1])),
+            inds=(mpo1.inds[-2], mpo2.inds[-2], '-1')
+        )
+        self._fused_mpo_cache.pop(locus + 1)
+        rand_id = qtn.rand_uuid()
+        self._fused_mpo_cache[locus] = qtn.TensorNetwork([
+            mpo1,
+            mpo2,
+            projector,
+            projector.copy().reindex({
+                mpo1.inds[-2]: mpo1.inds[-1],
+                mpo2.inds[-2]: mpo2.inds[-1],
+                '-1': '-2'
+            })
+        ]).contract().reindex({'-1': f'k{rand_id}', '-2': f'b{rand_id}'})
+        return projector.data
 
-    def neighbouring_bonds(self, bond: int) -> List[int]:
-        if bond == 0:
-            neighbours = [bond]
-        elif bond == self.n_nodes - 1:
-            neighbours = [bond - 1]
-        else:
-            neighbours = [bond - 1, bond]
-        return neighbours
-
-    def run(self) -> None:
-        """
-
-        Returns:
-
-        """
-        evals = self.gap_cache.evals if self.n_nodes == 2 else None
+    def run(self):
         for step in count(start=1):
-            max_gapped_bond = np.argmax(np.array(self.gap_cache.gap))
-            V, W = self.spectrum_projector(max_gapped_bond, self.gap_cache.evecs[max_gapped_bond])
-            horizon = self.tree.horizon
-            logging.info(f"step {step}, merging TreeNode({horizon[max_gapped_bond]}) "
-                         f"and TreeNode({horizon[max_gapped_bond + 1]}) to TreeNode({self.N + step - 1})")
-            self._tree.append(
-                TreeNode(
-                    id=self.N + step - 1,
-                    node=V,
-                    left=self.tree.node(horizon[max_gapped_bond]),
-                    right=self.tree.node(horizon[max_gapped_bond + 1])
-                )
+            max_gapped_bond = int(np.argmax(np.array(self._gap_cache.gap)))
+            logging.info(f"On step {step}, "
+                         f"merging Node({self.tree.horizon[max_gapped_bond]}) "
+                         f"and Node({self.tree.horizon[max_gapped_bond + 1]}) "
+                         f"into Node({self.n_sites + step - 1}).")
+            self.tree.fuse(
+                left_id=self.tree.horizon[max_gapped_bond],
+                right_id=self.tree.horizon[max_gapped_bond + 1],
+                new_id=self.n_sites + step - 1,
+                data=self.spectrum_projector(max_gapped_bond, self._gap_cache.evecs[max_gapped_bond])
             )
-            self.mpo.nodes.pop(max_gapped_bond)
-            self.mpo.nodes[max_gapped_bond] = W
-            self.gap_cache.gap.pop(max_gapped_bond)
-            self.gap_cache.evecs.pop(max_gapped_bond)
-            if self.n_nodes == 1:
-                logging.info('Reach the root in tree')
-                logging.info(f"Obtain ground state energies {evals}")
-                assert step == self.N - 1, "step out of range"
+            if len(self._fused_mpo_cache) == 1:
+                logging.info('Reaching the root in tree.')
+                logging.info(f"Obtain ground state energies, {repr(self.evals)}.")
+                assert step == self.n_sites - 1, "step is out of range."
                 break
-            for bond in self.neighbouring_bonds(max_gapped_bond):
-                evals, evecs = self.eigen_solver(self.block_hamiltonian(bond))
-                self.gap_cache.gap[bond] = self.truncation_gap(evals)
-                self.gap_cache.evecs[bond] = evecs
+            self._gap_cache.update(max_gapped_bond)
 
-    def reduced_density_matrix(self, on_site: int, energy_level: int = 0) -> Union[Tensor, np.ndarray]:
-        assert 0 <= on_site < self.N
-        assert 0 <= energy_level < self.physical_dimensions ** self.chi
+    @property
+    def measurements(self) -> 'TreeTensorNetworkMeasurements':
+        if not len(self._fused_mpo_cache) == 1:
+            raise RuntimeError("tSDRG algorithm hasn't been executed yet.")
+        return TreeTensorNetworkMeasurements(self._tree)
 
-        def descendant_chirality(ancestor: int, descendant: int) -> Tuple[int, str]:
-            path = self.tree.find_path(descendant)
-            return (ancestor, 'left') if self.tree.node(ancestor).left.id in path else (ancestor, 'right')
-        iterator = range(on_site+1) if (on_site + 1) / (self.N - on_site) < 1 else range(on_site, self.N)
-        open_bonds = [
-            descendant_chirality(
-                self.tree.lowest_common_ancestor(node_id, on_site),
-                node_id
-            ) for node_id in iterator
-        ]
-        rho = self.tree.contract_nodes(
-            self.tree.find_path(on_site)[:-1],
-            open_bonds=list(set(open_bonds))
-        ).reshape((
-            min(self.physical_dimensions ** self.N, self.chi),
-            self.physical_dimensions ** min(on_site + 1, self.N - on_site),
-            min(self.physical_dimensions ** self.N, self.chi),
-            self.physical_dimensions ** min(on_site + 1, self.N - on_site)
-        ))
-        return rho[energy_level, :, energy_level, :]
 
-    def entanglement_entropy(self, on_site: int, energy_level: int = 0) -> float:
-        ss = np.linalg.svd(self.reduced_density_matrix(on_site, energy_level), compute_uv=False)
-        return -1 * ss @ np.log(ss)
+class TreeTensorNetworkMeasurements:
 
-    def correlation_function(self):
+    def __init__(self, tree: TensorTree):
+        self._tree = tree
+
+    @property
+    def tree(self):
+        return self._tree
+
+    def sandwich(self, mpo: MatrixProductOperator = None) -> qtn.Tensor:
+        ket = self.tree.tensor_network()
+        bra = self.tree.tensor_network(conj=True, mangle_outer=False) if mpo is None \
+            else self.tree.tensor_network(conj=True)
+        net = (bra & ket) if mpo is None else (bra & mpo & ket)
+        return net.contract(all)
+
+    def expectation_value(self, mpo: MatrixProductOperator) -> np.ndarray:
+        exp_val = self.sandwich(mpo).data
+        if not np.allclose(np.zeros(exp_val.shape), exp_val - np.diagflat(np.diag(exp_val)), atol=1e-12):
+            logging.warning("Expectation value may contain large off-diagonal elements.")
+        return np.diag(exp_val)
+
+    def _min_surface(self, bipartite_site: int) -> Tuple[Dict, Dict]:
+        min_side = 0 if (bipartite_site + 1) / (self.tree.n_leaves - bipartite_site) < 1 else 1
+        iterator = range(bipartite_site + 1) if min_side == 0 \
+            else range(bipartite_site, self.tree.n_leaves)
+        ket_inds_map, bra_inds_map = {}, {}
+        for leaf_id in iterator:
+            lca_id = self.tree.common_ancestor(leaf_id, bipartite_site + 1, lowest=True)
+            ket_inds_map[f'{self.tree[lca_id].inds[min_side]}'] = f'rho_k{leaf_id}'
+            bra_inds_map[f'{self.tree[lca_id].inds[min_side]}'] = f'rho_b{leaf_id}'
+        return ket_inds_map, bra_inds_map
+
+    def reduced_density_matrix(self, site: int, level_idx: int) -> np.ndarray:
+        if not 0 <= site < self.tree.n_leaves - 1:
+            raise ValueError("Parameter `site` for bi-partition has to be within the system size.")
+        if not 0 <= level_idx < self.tree.root.shape[2]:
+            raise ValueError("Parameter `level_idx` has to be lower than truncation dimension.")
+        node_ids = self.tree.find_path(site + 1)
+        on_min_ket_surface, on_min_bra_surface = self._min_surface(site)
+        ket = self.tree.tensor_network(node_ids).reindex(on_min_ket_surface)
+        ket.isel(
+            {f'{TensorTree.Syntax.node}{self.tree.root_id}{TensorTree.Syntax.level_idx}': level_idx},
+            inplace=True
+        )
+        bra = self.tree.tensor_network(node_ids, conj=True, mangle_outer=False).reindex(on_min_bra_surface)
+        bra.isel(
+            {f'{TensorTree.Syntax.conj_node}{self.tree.root_id}{TensorTree.Syntax.level_idx}': level_idx},
+            inplace=True
+        )
+        return (ket & bra).contract(
+            [elem for node_id in node_ids
+             for elem in (f'{TensorTree.Syntax.node}{node_id}', f'{TensorTree.Syntax.conj_node}{node_id}')]
+        ).to_dense([*on_min_ket_surface.values()], [*on_min_bra_surface.values()])
+
+    def entanglement_entropy(self, site: int, level_idx: int) -> float:
+        rho = np.linalg.eigvalsh(self.reduced_density_matrix(site, level_idx))[::-1]
+        return -1 * rho @ np.log(rho)
+
+    def one_point_function(self):
         NotImplemented
 
-    def energies(self, mpo: MPO = None) -> Union[Tensor, np.ndarray]:
-        node_queue = deque([self.tree.root])
-        node_ids = self.tree.filter_out_leaf(self.tree.keys)
-        contractible = self.tree.Contractible(self.tree, node_ids, [])
-
-        def update_contractible(parent: TreeNode, on: str):
-            child = getattr(parent, on)
-            on_bond_idx = 0 if on == 'left' else 1
-            if child.is_leaf:
-                contractible[f"{parent.id}"][on_bond_idx] = f"NodeToMPO{child.id}"
-                contractible[f"conj{parent.id}"][on_bond_idx] = f"ConjNodeToMPO{child.id}"
-            else:
-                contractible[f"{parent.id}"][on_bond_idx] = f"Node{parent.id}ToNode{child.id}"
-                contractible[f"{child.id}"][2] = f"Node{parent.id}ToNode{child.id}"
-                contractible[f"conj{parent.id}"][on_bond_idx] = f"ConjNode{parent.id}ToConjNode{child.id}"
-                contractible[f"conj{child.id}"][2] = f"ConjNode{parent.id}ToConjNode{child.id}"
-                node_queue.append(child)
-
-        def mpo_network(node_id: int):
-            net = (
-                f"MPO{node_id - 1}ToMPO{node_id}",
-                f"MPO{node_id}ToMPO{node_id + 1}",
-                f"NodeToMPO{node_id}",
-                f"ConjNodeToMPO{node_id}"
-            )
-            if node_id == 0:
-                return net[1:]
-            elif node_id == self.N - 1:
-                return (net[0], *net[2:])
-            return net
-
-        while len(node_queue) > 0:
-            current_node = node_queue.popleft()
-            update_contractible(current_node, on='left')
-            update_contractible(current_node, on='right')
-
-        mpo_tensors = [node.tensor for node in mpo] if mpo is not None \
-            else [tree_node.node.tensor for tree_node in self.tree.leaves.values()]
-        mpo_network_structure = [mpo_network(node_id) for node_id in self.tree.leaves.keys()]
-
-        # TODO: con_order should be given
-        return ncon(
-            contractible.tensors + mpo_tensors,
-            contractible.network_structure + mpo_network_structure,
-            out_order=contractible.out_order
-        )
+    def two_point_function(self):
+        NotImplemented
